refactor: unify icons folder
