+++ src/bricks/icons/icons.tsx
@@ -1,0 +1,69 @@
+import * as React from 'react';
+
+export function EditIcon(): JSX.Element {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+    >
+      <path d="M3,3 L6,3 L6,5 L3,5 C2.44771525,5 2,5.44771525 2,6 L2,13 C2,13.5522847 2.44771525,14 3,14 L10,14 C10.5522847,14 11,13.5522847 11,13 L11,10 L13,10 L13,13 C13,14.6568542 11.6568542,16 10,16 L3,16 C1.34314575,16 0,14.6568542 0,13 L0,6 C0,4.34314575 1.34314575,3 3,3 Z M11,2 L14,5 L7,12 L4,12 L4,9 L11,2 Z M12,1 C13.2893219,-0.201010127 14.5929291,-0.201010127 15,1 C16.2010101,1.40707089 16.2010101,2.71067812 15,4 L12,1 Z" />
+    </svg>
+  );
+}
+
+export function Logo(props: {
+  width?: string;
+  height?: string;
+  style?: React.CSSProperties;
+}): JSX.Element {
+  return (
+    <svg
+      style={props.style}
+      xmlns="http://www.w3.org/2000/svg"
+      width={props.width}
+      height={props.height}
+      viewBox="0 0 25 21"
+    >
+      <path d="M6.88538031,12.3145836 C7.5626171,11.9709698 8.38096193,11.5620105 9.39258852,11.0611807 C9.32139178,8.04799842 8.96035589,4.35591876 9.20846395,1.1105511 C9.20846395,0.603565145 9.40438871,0.0575802729 9.79623824,0.0575802729 C10.5407524,0.0575802729 11.4028213,0.525567306 11.4028213,0.876557581 C11.4028213,1.53953921 11.2068966,3.17749383 11.2068966,6.72639549 C11.2068966,8.29267717 11.2455408,9.29303652 11.2938192,10.1234363 C11.8127299,9.86823589 12.370273,9.59450422 12.9702194,9.30032417 C13.2832652,9.1459353 13.5852957,8.99695347 13.8767341,8.85316284 C13.8542036,5.84480386 13.8401966,2.65346644 13.9106583,1.22754786 C13.9106583,0.720561903 14.2633229,0.291573789 14.6551724,0.291573789 C15.3996865,0.291573789 16.2225705,0.759560822 16.2225705,1.1105511 C16.2225705,3.19885916 15.9143186,4.48184751 15.9091565,7.84960276 C16.7571551,7.43042436 17.4865063,7.06912465 18.1158484,6.75619478 C18.1803757,4.96372762 18.2465501,3.13371976 18.299373,1.18854894 C18.299373,-0.839394874 20.3369906,-0.0594164854 20.3369906,1.9295284 C20.3369906,2.72599699 20.3011752,4.25913555 20.2690824,5.67635654 C21.9753115,4.80813626 22.2054762,4.63945352 22.6489028,4.30846249 C23.1583072,3.9184733 23.354232,3.64548086 23.6285266,3.64548086 C24.5689655,3.64548086 25,4.58145493 25,5.20543764 C25,5.63442575 24.7257053,5.86841927 24.177116,6.18041062 C23.354232,6.68739657 22.3746082,7.07738577 21.7868339,7.11638469 C21.2451687,7.14333937 20.843896,7.27741725 20.2288405,7.59909638 C20.2229811,7.94245807 20.2194357,8.23239326 20.2194357,8.44234795 L20.2194357,18.6210659 C20.2194357,19.2840475 19.200627,20.2200216 18.7304075,20.2200216 C18.0642633,20.2200216 17.8683386,18.075081 17.8683386,17.4900973 C17.8683386,14.421748 17.9458558,11.639953 18.0420208,8.84004395 C17.4602276,9.16993946 16.7621495,9.55803056 15.9090909,10.0131099 C15.9090909,11.4822447 15.9090909,12.9411953 15.9090909,14.2921859 C15.9090909,17.2171048 16.4184953,17.139107 16.4184953,17.607094 C16.4184953,18.0360821 15.6739812,20.2200216 14.8902821,20.2200216 C14.2241379,20.2200216 14.2241379,19.55704 14.2241379,18.8940584 C14.2241379,17.9970832 13.9106583,15.9301405 13.9106583,14.0581923 C13.9106583,13.2698308 13.9033365,12.2239949 13.8941748,11.0613229 C13.3245111,11.3508021 12.7049027,11.6603558 12.0297806,11.9912496 C11.8243374,12.0924403 11.6240533,12.1916218 11.428782,12.2888427 C11.4597153,12.9242271 11.4811912,13.667324 11.4811912,14.6821751 C11.4811912,16.632121 11.4028213,18.7770616 11.4028213,19.1670508 C11.4028213,19.7130357 10.6191223,21 9.9137931,21 C9.2476489,21 9.01253918,20.2590205 8.93416928,19.3230465 C8.93416928,17.38571 9.32398841,15.1089013 9.41202253,13.3147809 C8.41289749,13.8355262 7.58912679,14.286387 6.90461945,14.6793365 L6.93573668,18.5040692 C6.93573668,19.3230465 6.54388715,20.4540151 5.56426332,20.4540151 C5.21159875,20.4540151 5.05485893,20.2980195 4.93730408,18.6600648 C4.87904543,17.8579797 4.84003493,16.9840568 4.81431063,15.9771794 C3.7421623,16.7164427 3.40949868,17.1150509 2.97805643,17.4510983 C1.05799373,19.0110551 0,17.9580843 0,17.4900973 C0,16.4371264 1.21473354,15.6181491 2.62539185,14.721174 C3.38068615,14.2356956 3.902187,13.901147 4.77603017,13.4195384 C4.75567325,10.5885996 4.78056426,6.82926813 4.78056426,1.42254245 C4.78056426,0.603565145 4.78056426,0.0575802729 5.99529781,0.0575802729 C6.54388715,0.0575802729 7.2492163,0.759560822 7.2492163,2.39751544 C7.2492163,3.64548086 6.85736677,7.7403674 6.85736677,8.87133606 L6.88538031,12.3145836 Z" />
+    </svg>
+  );
+}
+
+export function TransactionIcon(): JSX.Element {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="18"
+      viewBox="0 0 16 18"
+    >
+      <path d="M11.6850112,6 L5,6 C3.34314575,6 2,7.34314575 2,9 L2,10 L0,10 L0,9 C0,6.23857625 2.23857625,4 5,4 L11.6850112,4 L10.6073147,3.49025739 L8.67276818,1.50221105 L10.1345536,0 L15,5 L10.1345536,10 L8.67276818,8.49778895 L10.6641888,6.45129568 L11.6850112,6 Z M4.31498875,12 L11,12 C12.6568542,12 14,10.6568542 14,9 L14,8 L16,8 L16,9 C16,11.7614237 13.7614237,14 11,14 L4.31498875,14 L5.39268526,14.5097426 L7.32723182,16.497789 L5.86544636,18 L1,13 L5.86544636,8 L7.32723182,9.50221105 L5.33581117,11.5487043 L4.31498875,12 Z" />
+    </svg>
+  );
+}
+
+export const ShoppingBagIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="15"
+    viewBox="0 0 16 15"
+  >
+    <path d="M1.15555556,7 L1,7 C0.44771525,7 0,6.55228475 0,6 C0,5.44771525 0.44771525,5 1,5 L3.05025253,5 L3,4.94974747 L7.24264069,0.707106781 C7.63316498,0.316582489 8.26632996,0.316582489 8.65685425,0.707106781 L12.8994949,4.94974747 L12.8492424,5 L15,5 C15.5522847,5 16,5.44771525 16,6 C16,6.55228475 15.5522847,7 15,7 L14.8444444,7 L13.8632912,13.307414 C13.711755,14.2815752 12.8729351,15 11.8870582,15 L4.11294178,15 C3.1270649,15 2.28824503,14.2815752 2.13670885,13.307414 L1.15555556,7 Z M10.4748737,5 L7.94974747,2.47487373 L5.4246212,5 L10.4748737,5 Z M5.64704761,8.08518543 C5.38031456,8.01371453 5.10614608,8.17200577 5.03467518,8.43873882 L3.999399,12.3024421 C3.92792809,12.5691752 4.08621934,12.8433437 4.35295239,12.9148146 C4.61968544,12.9862855 4.89385392,12.8279942 4.96532482,12.5612612 L6.000601,8.69755787 C6.07207191,8.43082482 5.91378066,8.15665634 5.64704761,8.08518543 Z M8.64704761,8.08518543 C8.38031456,8.01371453 8.10614608,8.17200577 8.03467518,8.43873882 L6.999399,12.3024421 C6.92792809,12.5691752 7.08621934,12.8433437 7.35295239,12.9148146 C7.61968544,12.9862855 7.89385392,12.8279942 7.96532482,12.5612612 L9.000601,8.69755787 C9.07207191,8.43082482 8.91378066,8.15665634 8.64704761,8.08518543 Z M11.6470476,8.08518543 C11.3803146,8.01371453 11.1061461,8.17200577 11.0346752,8.43873882 L9.999399,12.3024421 C9.92792809,12.5691752 10.0862193,12.8433437 10.3529524,12.9148146 C10.6196854,12.9862855 10.8938539,12.8279942 10.9653248,12.5612612 L12.000601,8.69755787 C12.0720719,8.43082482 11.9137807,8.15665634 11.6470476,8.08518543 Z" />
+  </svg>
+);
+
+export function ProductIcon(): JSX.Element {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+    >
+      <path d="M1,0 L6,0 C6.55228475,0 7,0.44771525 7,1 L7,6 C7,6.55228475 6.55228475,7 6,7 L1,7 C0.44771525,7 0,6.55228475 0,6 L0,1 C0,0.44771525 0.44771525,0 1,0 Z M10,0 L15,0 C15.5522847,0 16,0.44771525 16,1 L16,6 C16,6.55228475 15.5522847,7 15,7 L10,7 C9.44771525,7 9,6.55228475 9,6 L9,1 C9,0.44771525 9.44771525,0 10,0 Z M12.5,9 C14.4329966,9 16,10.5670034 16,12.5 C16,14.4329966 14.4329966,16 12.5,16 C10.5670034,16 9,14.4329966 9,12.5 C9,10.5670034 10.5670034,9 12.5,9 Z M1,9 L6,9 C6.55228475,9 7,9.44771525 7,10 L7,15 C7,15.5522847 6.55228475,16 6,16 L1,16 C0.44771525,16 0,15.5522847 0,15 L0,10 C0,9.44771525 0.44771525,9 1,9 Z" />
+    </svg>
+  );
+}
