+++ src/bricks/icons/index.ts
@@ -6,3 +6,5 @@ export * from './plus';
 export * from './paypal';
 export * from './search-minus';
 export * from './search-plus';
+export * from './search';
+export * from './icons';
