+++ src/bricks/icons/search.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
 
-export function SearchIcon(): JSX.Element {
+export function SearchIcon({ style }: any): JSX.Element {
   return (
     <svg
+      style={style}
       xmlns="http://www.w3.org/2000/svg"
       width="15"
       height="15"
