+++ src/components/settings/scaling-buttons.tsx
@@ -3,14 +3,20 @@ import { useLocalStorage } from '../../hooks/use-storage';
 import { Button, SearchMinus, SearchPlus } from '../../bricks';
 
 export const useScalingState = () => {
-  const [scaling, setScaling] = useLocalStorage('scaling', 0);
+  const [scaling, setScaling] = useLocalStorage('scaling', 16);
   const increment = () => {
     setScaling(scaling + 1);
-    window.location.reload();
+    document.documentElement.style.setProperty(
+      '--baseFontSize',
+      `${scaling}px`
+    );
   };
   const decrement = () => {
     setScaling(scaling - 1);
-    window.location.reload();
+    document.documentElement.style.setProperty(
+      '--baseFontSize',
+      `${scaling}px`
+    );
   };
 
   return {
