--- src/components/ui/icons/add.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-
-export interface AddIconProps {
-  isPrimary?: boolean;
-}
-
-function Icon(): JSX.Element {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-    >
-      <polygon points="9 7 16 7 16 9 9 9 9 16 7 16 7 9 0 9 0 7 7 7 7 0 9 0" />
-    </svg>
-  );
-}
-
-export const AddIcon = Icon;
-export const CancelIcon = Icon;
