--- src/components/ui/icons/arrow-down.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-
-export function ArrowDownIcon(): JSX.Element {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="14"
-      height="8"
-      viewBox="0 0 14 8"
-    >
-      <polygon points="7 6 7.5 5 12.5 0 14 1.5 7.5 8 6.5 8 0 1.5 1.5 0 6.5 5" />
-    </svg>
-  );
-}
