--- src/components/ui/icons/logo.tsx
@@ -1,19 +1,0 @@
-import * as React from 'react';
-
-export function Logo(props: {
-  width?: string;
-  height?: string;
-  style?: React.CSSProperties;
-}): JSX.Element {
-  return (
-    <svg
-      style={props.style}
-      xmlns="http://www.w3.org/2000/svg"
-      width={props.width}
-      height={props.height}
-      viewBox="0 0 25 21"
-    >
-      <path d="M6.88538031,12.3145836 C7.5626171,11.9709698 8.38096193,11.5620105 9.39258852,11.0611807 C9.32139178,8.04799842 8.96035589,4.35591876 9.20846395,1.1105511 C9.20846395,0.603565145 9.40438871,0.0575802729 9.79623824,0.0575802729 C10.5407524,0.0575802729 11.4028213,0.525567306 11.4028213,0.876557581 C11.4028213,1.53953921 11.2068966,3.17749383 11.2068966,6.72639549 C11.2068966,8.29267717 11.2455408,9.29303652 11.2938192,10.1234363 C11.8127299,9.86823589 12.370273,9.59450422 12.9702194,9.30032417 C13.2832652,9.1459353 13.5852957,8.99695347 13.8767341,8.85316284 C13.8542036,5.84480386 13.8401966,2.65346644 13.9106583,1.22754786 C13.9106583,0.720561903 14.2633229,0.291573789 14.6551724,0.291573789 C15.3996865,0.291573789 16.2225705,0.759560822 16.2225705,1.1105511 C16.2225705,3.19885916 15.9143186,4.48184751 15.9091565,7.84960276 C16.7571551,7.43042436 17.4865063,7.06912465 18.1158484,6.75619478 C18.1803757,4.96372762 18.2465501,3.13371976 18.299373,1.18854894 C18.299373,-0.839394874 20.3369906,-0.0594164854 20.3369906,1.9295284 C20.3369906,2.72599699 20.3011752,4.25913555 20.2690824,5.67635654 C21.9753115,4.80813626 22.2054762,4.63945352 22.6489028,4.30846249 C23.1583072,3.9184733 23.354232,3.64548086 23.6285266,3.64548086 C24.5689655,3.64548086 25,4.58145493 25,5.20543764 C25,5.63442575 24.7257053,5.86841927 24.177116,6.18041062 C23.354232,6.68739657 22.3746082,7.07738577 21.7868339,7.11638469 C21.2451687,7.14333937 20.843896,7.27741725 20.2288405,7.59909638 C20.2229811,7.94245807 20.2194357,8.23239326 20.2194357,8.44234795 L20.2194357,18.6210659 C20.2194357,19.2840475 19.200627,20.2200216 18.7304075,20.2200216 C18.0642633,20.2200216 17.8683386,18.075081 17.8683386,17.4900973 C17.8683386,14.421748 17.9458558,11.639953 18.0420208,8.84004395 C17.4602276,9.16993946 16.7621495,9.55803056 15.9090909,10.0131099 C15.9090909,11.4822447 15.9090909,12.9411953 15.9090909,14.2921859 C15.9090909,17.2171048 16.4184953,17.139107 16.4184953,17.607094 C16.4184953,18.0360821 15.6739812,20.2200216 14.8902821,20.2200216 C14.2241379,20.2200216 14.2241379,19.55704 14.2241379,18.8940584 C14.2241379,17.9970832 13.9106583,15.9301405 13.9106583,14.0581923 C13.9106583,13.2698308 13.9033365,12.2239949 13.8941748,11.0613229 C13.3245111,11.3508021 12.7049027,11.6603558 12.0297806,11.9912496 C11.8243374,12.0924403 11.6240533,12.1916218 11.428782,12.2888427 C11.4597153,12.9242271 11.4811912,13.667324 11.4811912,14.6821751 C11.4811912,16.632121 11.4028213,18.7770616 11.4028213,19.1670508 C11.4028213,19.7130357 10.6191223,21 9.9137931,21 C9.2476489,21 9.01253918,20.2590205 8.93416928,19.3230465 C8.93416928,17.38571 9.32398841,15.1089013 9.41202253,13.3147809 C8.41289749,13.8355262 7.58912679,14.286387 6.90461945,14.6793365 L6.93573668,18.5040692 C6.93573668,19.3230465 6.54388715,20.4540151 5.56426332,20.4540151 C5.21159875,20.4540151 5.05485893,20.2980195 4.93730408,18.6600648 C4.87904543,17.8579797 4.84003493,16.9840568 4.81431063,15.9771794 C3.7421623,16.7164427 3.40949868,17.1150509 2.97805643,17.4510983 C1.05799373,19.0110551 0,17.9580843 0,17.4900973 C0,16.4371264 1.21473354,15.6181491 2.62539185,14.721174 C3.38068615,14.2356956 3.902187,13.901147 4.77603017,13.4195384 C4.75567325,10.5885996 4.78056426,6.82926813 4.78056426,1.42254245 C4.78056426,0.603565145 4.78056426,0.0575802729 5.99529781,0.0575802729 C6.54388715,0.0575802729 7.2492163,0.759560822 7.2492163,2.39751544 C7.2492163,3.64548086 6.85736677,7.7403674 6.85736677,8.87133606 L6.88538031,12.3145836 Z" />
-    </svg>
-  );
-}
